fix a bug in VIA emulation: latch will be cleared only, when reading IRA or reseting second bit in IFR register

+++ emulation/libc64/disk/via/via.h
@@ -85,7 +85,7 @@ protected:
     bool timerAToggle;
     
     uint16_t timerBCounter;
-    uint16_t timerBLatch;
+    uint8_t timerBLatch;
     uint16_t timerBCounterRead;
     bool timerBTrigger;
